[libs] E: Move Binding for open()
This commits moves the binding for the open() system call to the syscall
crate.
